Aggiunto classe singleton che gestisce l'utente loggato
