+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             android:label="LoginActivity"
             android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity
-            android:name=".Main2Activity"
+            android:name=".EsploraActivity"
             android:label="@string/title_activity_main2"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
