+++ app/src/main/res/layout/activity_main2.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Main2Activity">
+    tools:context=".EsploraActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -37,7 +37,7 @@
         android:layout_height="match_parent"
         android:scrollbars="vertical"
         android:id="@+id/recyclerView"
-        tools:context=".Main2Activity">
+        tools:context=".EsploraActivity">
 
     </android.support.v7.widget.RecyclerView>
 </android.support.design.widget.CoordinatorLayout>
