+++ app/src/main/res/layout/content_main2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".Main2Activity"
+    tools:context=".EsploraActivity"
     tools:showIn="@layout/activity_main2">
 
 </android.support.constraint.ConstraintLayout>
